Add a filter to $url in init()
By allowing developers to filter the $url, we can customize our installation and avoid complicated regex rules. My sites never care about query variables but we always have to account for them in our redirections with complicated regex statements. By filtering the $url, we can chop off the query part of a URL and never have to deal with it in the redirection rule.

I initially proposed this change and gave examples here https://wordpress.org/support/topic/making-it-easy-to-redirect-without-regex