Do not wrap encode_from in ^ and $ if regex
If the from url is a regex then we should not wrap the encoded string with beginning and end regex characters.

Note: I have not tested whether we should encode regex strings or not - but I suspect not. This needs knowledge about url encoding that I don't have in my head and don't have time to research and a decent set of test cases and testing.