Allow user maintenance of item Position field. The admin UI has been enhanced to allow edit of the Position field which is used to determine the redirection item processing order. As a part of this effort the UI also now displays the item description.
The first problem solved with this is that after creating redirection items it was difficult to know the order in which they will be executed without opening a tool like phpMyAdmin. The second problem solved is that the only way to change the processing order was to manually poke the DBMS, or to delete and recreate the items in the UI.

Screenshot position_mods.png is provided of the UI update. The UI description is prefaced with the position in the same format as calculated internally: group*1000+position. So a redirection rule assigned position 10 will be shown with the prefix 1010, 2010, 3010, etc., depending on the group to which it has been assigned.
To change the order in which redirects are processed, go to item detail and use the spinner or manually enter a number from 0 to 999 in the numeric-only field.
The list does not re-sort automatically since only part of the page is ajaxified. To refresh the list in the new order, refresh the page or click the URL column header.

Note 1: The column header has been temporarily frozen to sort by-ascending on the modified position_desc_source field, though the column header still says "URL". There is no sense in sorting the list in descending order. In another update, if required, the column header can be renamed to Description and the position might be separated into a new display column so that the fields can be sorted independently.

New 2: Moving a redirection item to a different group changes the order in which it's processed. This was not obvious before but is now. This might help users to understand and better manage their rules.

This change was made to production build 2.6.5 and merged with the current master version. The three files were directly modified. If these are generated in a build process the code will need to be extracted out to the correct source.