Client shows itself instead of waiting for page load
The React code shows itself when fully loaded. It has no other dependencies, other than HTML on the page and so we push loading of the file to the bottom of the page.

This should get around errors caused by other JS loading and showing an error that then stops jQuery.ready